rgb colour codes to plot 2015 National Land Cover.

A dataset containing red, blue, green values associated with all the land cover types found in the 2015 Canadian National Land Cover Classification

Usage

clrfile

Format

A data frame with 20 rows and 6 variables:

Value

price, in US dollars

red

weight of the diamond, in carats

green

weight of the diamond, in carats

blue

weight of the diamond, in carats

rgb

weight of the diamond, in carats

LCC_NAME

weight of the diamond, in carats

Source

https://open.canada.ca/data/en/dataset/4e615eae-b90c-420b-adee-2ca35896caf6