Variables for cost estimation

Cost variables for estimate_cost_study_area

Usage

cost_vars

Format

A list with 18 variables used to calculate the cost model for N Ontario:

truck_cost_per_day

Cost of truck use per day per crew, in dollars

truck_n_crews

Number of crews for truck surveys

truck_arus_per_crew_per_day

number of arus deployed per crew per day

atv_cost_per_day

Cost of ATV use per day per crew, in dollars

atv_n_crews

Number of crews for ATV surveys

atv_arus_per_crew_per_day

Number of arus deployed per crew per day

helicopter_cost_per_hour

Cost of helicopter rental per hour, in dollars

helicopter_max_km_from_base

Maximum range of helicopter from fuelling base, in kilometres

helicopter_base_setup_cost_per_km

Cost of setting up base for helicopter use with distance from airport, in km

helicopter_l_per_hour

Helicopter fuel usage per hour, in litres

helicopter_crew_size

Helicopter crew size

helicopter_aru_per_person_per_day

Number of arus deployed per day per person

helicopter_relocation_speed

Speed of movement of helicopter when relocating, in km per hour

helicopter_airport_cost_per_l

Cost of heli fuel from an airport, in dollars per litre

helicopter_base_cost_per_l

Cost of heli fuel from a basecamp, in dollars per litre

helicopter_2nd_base_cost_per_l

Cost of heli fuel from a a remote fuel cache, in dollars per litre

helicopter_hours_flying_within_sa_per_day

Number of hours a helicopter spends flying in the study area per day.

...

Source

Advice from Rich Russell